9.2 webpack part2 error

+++ src/controllers/videoControllers.js
@@ -1,4 +1,5 @@
 import { query } from "express";
+import User from "../models/User";
 import Video from "../models/video";
 
 /*
@@ -23,7 +24,7 @@ export const home = async (req, res) => {
 };
 export const watch = async (req, res) => {
   const { id } = req.params; //const id = req.params.id
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
@@ -31,20 +32,35 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
   const { id } = req.params; //const id = req.params.id
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id); // return VIdeo model
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
-  return res.render("edit", { pageTitle: `Editing: ${video.title} `, video });
+  if (String(video.owner) !== String(req.session.user._id)) {
+    return res.status(403).redirect("/");
+  }
+  return res.render("edit", {
+    pageTitle: `Editing: ${video.title} `,
+    video,
+  });
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params; //const id = req.params.id
   const { title, description, hashtags } = req.body;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.exists({ _id: id }); // return true or false
 
   if (!video) {
     console.log("post Edit");
     return res.render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(req.session.user._id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -59,14 +75,22 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
-  console.log(description);
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    await user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -78,6 +102,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    console.log("post Edit");
+    return res.render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(req.session.user._id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id); //same findOneAndElete({_id:id})
   return res.redirect("/");
 };
